Split out aprs messaging code from aprs.c.
