+++ aprs.c
@@ -12,7 +12,6 @@
 #include <sys/stat.h>
 #include <sys/utsname.h>
 #include <unistd.h>
-#include <time.h>
 #include <math.h>
 #include <errno.h>
 #include <termios.h>
@@ -20,7 +19,6 @@
 #include <netdb.h>
 #include <ctype.h>
 #include <json/json.h>
-#include <assert.h>
 #include <netinet/if_ether.h> /* only for ETH_P_AX25, canned packets */
 
 #include "util.h"
@@ -29,6 +27,7 @@
 #include "ui.h"
 #include "aprs-is.h"
 #include "aprs-ax25.h"
+#include "aprs-msg.h"
 
 #ifndef BUILD
 #define BUILD 0
@@ -38,12 +37,6 @@
 #define REVISION "Unknown"
 #endif
 
-
-#define OBJNAME(p) ((p)->object_or_item_name ? (p)->object_or_item_name : \
-                    (p)->src_callsign)
-
-int store_packet(struct state *state, fap_packet_t *fap);
-int _ui_send(struct state *state, const char *name, const char *value);
 
 void fap_free_wrapper(char *dispStr,  struct state *state, fap_packet_t *fap)
 {
@@ -59,40 +52,17 @@ void display_fap_error(char *string, struct state *state, fap_packet_t *fap)
         char errbuf[1024];
 
         fap_explain_error(*fap->error_code, errbuf);
-        printf("FAP Error %s: %s: %i: %s\n",
-               time2str(fap->timestamp, 0), string, *fap->error_code, errbuf);
+        printf("\n%s FAP Error: %s: %i: %s\n",
+               time2str(fap->timestamp, 0),
+               string,
+               *fap->error_code,
+               errbuf);
+
+        if(state->debug.console_display_filter & CONSOLE_DISPLAY_FAPERR) {
+                printf("Error Pkt: %s\n", fap->orig_packet);
+        }
 
         state->stats.fapErrCount++;
-}
-
-/*
- * Debug only - look at message packets
- */
-void display_fap_message(struct state *state, fap_packet_t *fap)
-{
-        state->stats.inMsgCount++;
-
-        printf("MESSAGE[%lu]: ", state->stats.inMsgCount);
-        if(fap->messaging) {
-                printf("msg capab: %s ", (*fap->messaging == 1) ? "yes" : "no");
-        }
-        if(fap->destination) {
-                printf("dest: %s ", fap->destination);
-        }
-        if(fap->message) {
-                printf("msg[%d]: %s ", strlen(fap->message), fap->message);
-        }
-        if(fap->message_ack) {
-                printf("ack: %s ", fap->message_ack);
-        }
-        if(fap->message_nack) {
-                printf("nack: %s ", fap->message_nack);
-        }
-        if(fap->message_id) {
-                printf("id: %s", fap->message_id);
-        }
-        printf("\n");
-        fflush(stdout);
 }
 
 void display_fap_pkt(char *string, fap_packet_t *fap)
@@ -113,7 +83,7 @@ void dump_ui_msg(struct ui_msg *ui_msg)
         int i;
         unsigned char *buf = (unsigned char *)ui_msg;
 
-        printf("ui message [%d]: ", sizeof(struct ui_msg));
+        printf("ui message [%zu]: ", sizeof(struct ui_msg));
 
         for(i=0; i < ui_msg->length; i++) {
                 printf("%02x ", *buf);
@@ -123,7 +93,7 @@ void dump_ui_msg(struct ui_msg *ui_msg)
 }
 
 
-int send_kiss_beacon(int fd, char *packet)
+bool send_kiss_beacon(int fd, char *packet)
 {
         char buf[512];
         int ret;
@@ -139,7 +109,7 @@ int send_kiss_beacon(int fd, char *packet)
         return write(fd, buf, len) == len;
 }
 
-int send_net_beacon(int fd, char *packet)
+bool send_net_beacon(int fd, char *packet)
 {
         int ret = 0;
 
@@ -153,10 +123,10 @@ int send_net_beacon(int fd, char *packet)
  * Send beacon packet out one of the interfaces serial, ax.25 or
  * network stack
  */
-int send_beacon(struct state *state, char *packet)
+bool send_beacon(struct state *state, char *packet)
 {
         if(state->debug.console_display_filter & CONSOLE_DISPLAY_PKTOUT) {
-                printf("\nPacket out %s: len: %d, src: %s, pkt: %s\n",
+                printf("\n%s Packet out: len: %zu, src: %s, pkt: %s\n",
                        time2str(NULL, 0), strlen(packet), state->ax25_srcaddr, packet);
         }
         if (STREQ(state->conf.tnc_type, "KISS")) {
@@ -174,31 +144,6 @@ int send_beacon(struct state *state, char *packet)
 
         } else {
                 return send_net_beacon(state->tncfd, packet);
-        }
-}
-
-/*
- * Send APRS message packet out one of the interfaces:
- * serial, ax.25 or network stack
- */
-int send_message(struct state *state, char *msg)
-{
-
-        if (STREQ(state->conf.tnc_type, "KISS")) {
-                _ui_send(state, "I_TX", "1000");
-                return send_kiss_beacon(state->tncfd, msg);
-        } else if (STREQ(state->conf.tnc_type, "AX25")) {
-#ifdef HAVE_AX25_TRUE
-                _ui_send(state, "I_TX", "1000");
-                return send_ax25_beacon(state, msg);
-#else
-                printf("%s:%s(): tnc type has been configured to use AX.25 but AX.25 lib has not been installed.\n",
-                       __FILE__, __FUNCTION__);
-                return 0;
-#endif /* #ifdef HAVE_AX25_TRUE */
-
-        } else {
-                return send_net_beacon(state->tncfd, msg);
         }
 }
 
@@ -581,7 +526,7 @@ void display_wx(struct state *state, fap_packet_t *fap)
                         state->last_wx = NULL;
                 }
                 if(state->debug.console_display_filter & CONSOLE_DISPLAY_WX) {
-                        printf("display_wx(%d): %s\n",
+                        printf("display_wx(%zu): %s\n",
                                strlen(fap->orig_packet), fap->orig_packet);
                 }
 
@@ -756,308 +701,6 @@ void display_dist_and_dir(struct state *state, fap_packet_t *fap)
         _ui_send(state, "AI_DISTANCE", buf);
 }
 
-#ifdef  WEBAPP
-char *format_msgtime(time_t *timestamp)
-{
-        static char timestr[32];
-        struct tm *tm;
-        time_t curtime;
-        time_t *thistime;
-
-        if(timestamp == NULL) {
-                curtime = time(NULL);
-                thistime=&curtime;
-        }else {
-                thistime=timestamp;
-        }
-        tm=localtime(thistime);
-
-        strftime(timestr, sizeof(timestr), "%H%M", tm);
-
-        return timestr;
-}
-
-bool isnewmessage(struct state *state, fap_packet_t *fap)
-{
-        aprsmsg_t *lastmsg;
-        int i;
-
-        /*
-         * Handle an ack
-         */
-        if(fap->message_ack != NULL) {
-                return(false);
-        }
-
-        for(i = 0; i < KEEP_MESSAGES; i++) {
-                if(state->msghist[i] != NULL) {
-
-                        lastmsg = state->msghist[i];
-
-#if 0
-                        printf("Message: Comparing these src calls: %s, %s, %d\n",
-                               lastmsg->srccall, OBJNAME(fap), STREQ(lastmsg->srccall, OBJNAME(fap) ) );
-
-                        printf("Message: Comparing these dst calls: %s, %s, %d\n",
-                               lastmsg->dstcall, fap->destination, STREQ(lastmsg->dstcall, fap->destination) );
-
-                        printf("Message: Comparing these msgs: %s, %s, %d\n",
-                               lastmsg->message, fap->message, STREQ(lastmsg->message, fap->message) );
-#endif
-
-                        /*
-                         * If the timestamps are ever compared need to add in
-                         * some slop for jitter.
-                         */
-                        if( STREQ(lastmsg->srccall, OBJNAME(fap) ) &&
-                            STREQ(lastmsg->dstcall, fap->destination) &&
-                            STREQ(lastmsg->message, fap->message) ) {
-                                return(false);
-                        }
-                }
-        }
-
-        return(true);
-}
-
-void save_message(struct state *state, fap_packet_t *fap)
-{
-        int msgind = state->msghist_idx;
-        char basecall[MAX_CALLSIGN];
-        int ssid;
-        aprsmsg_t *aprsmsg;
-        char aprs_ack[24]; /* largest APRS Message Ack is 19 bytes */
-
-        assert(msgind < KEEP_MESSAGES && msgind >= 0);
-
-        if(state->msghist[msgind] != NULL) {
-                aprsmsg = state->msghist[msgind];
-                free(aprsmsg->message);
-                if (aprsmsg->message_id != NULL) {
-                        free(aprsmsg->message_id);
-                }
-        } else {
-                if( (state->msghist[msgind] = malloc(sizeof(aprsmsg_t))) == NULL) {
-                        printf("%s: malloc message struct error: %s\n",
-                               __FUNCTION__, strerror(errno));
-                        return;
-                }
-                aprsmsg = state->msghist[msgind];
-        }
-
-        if( (aprsmsg->message = malloc(strlen(fap->message) + 1)) == NULL){
-                printf("%s: malloc message error: %s\n",
-                       __FUNCTION__, strerror(errno));
-                free(aprsmsg);
-                return;
-        }
-
-        if(fap->timestamp == NULL) {
-                aprsmsg->timestamp = time(NULL);
-        } else {
-                aprsmsg->timestamp = *fap->timestamp;
-        }
-
-        strcpy(aprsmsg->message, fap->message);
-        strncpy(aprsmsg->srccall, fap->src_callsign, MAX_CALLSIGN);
-        strncpy(aprsmsg->dstcall, fap->destination, MAX_CALLSIGN);
-
-        printf("Message: storing this src call: %s, check: %s\n",
-               aprsmsg->srccall, OBJNAME(fap) );
-
-        /*
-         * Handle message sent with an ID which is an ACK request
-         */
-        if(fap->message_id != NULL) {
-                if( (aprsmsg->message_id=calloc(MAX_MSGID + 1, 1)) == NULL) {
-                        printf("%s: malloc message id error: %s\n",
-                               __FUNCTION__, strerror(errno));
-                        return;
-                }
-
-                strncpy(aprsmsg->message_id, fap->message_id, MAX_MSGID);
-                get_base_callsign(basecall, &ssid, aprsmsg->dstcall);
-                printf("callsign check: %s %s\n", basecall, aprsmsg->dstcall);
-                if(STREQ(basecall, state->basecall)) {
-                        /* send an ack */
-                        pr_debug("SEND an ack to %s\n", fap->src_callsign);
-                        sprintf(aprs_ack,":%-9s:ack%s",
-                                  aprsmsg->dstcall,
-                                  aprsmsg->message_id);
-
-                        send_message(state, aprs_ack);
-                }
-        } else {
-                aprsmsg->message_id = NULL;
-        }
-
-        printf("save_message index: %d, from: %s, to: %s, ack: %s, id: %s, Msg: %s\n",
-               msgind,
-               fap->src_callsign, fap->destination,
-               (fap->message_ack == NULL) ? "None" : fap->message_ack,
-               (fap->message_id == NULL) ? "None" : fap->message_id,
-               fap->message);
-
-
-        state->msghist_idx = (state->msghist_idx == KEEP_MESSAGES - 1) ? 0 : ++msgind;
-        assert(state->msghist_idx < KEEP_MESSAGES && state->msghist_idx >= 0);
-
-}
-
-void webdisplay_message(struct state *state, fap_packet_t *fap)
-{
-        char buf[512];
-
-        printf("webdisplay_message: from: %s, to: %s, ack: %s, nack: %s, id: %s, Msg: %s\n",
-               fap->src_callsign, fap->destination,
-               (fap->message_ack == NULL) ? "None" : fap->message_ack,
-               (fap->message_nack == NULL) ? "None" : fap->message_nack,
-               (fap->message_id == NULL) ? "None" : fap->message_id,
-               fap->message);
-
-        snprintf(buf, sizeof(buf), "%s->%s@%s:%s",
-                 fap->src_callsign,
-                 fap->destination,
-                 format_msgtime(fap->timestamp),
-                 fap->message);
-
-        _ui_send(state, "MS_MESSAGE", buf);
-}
-
-/*
- * Copy a string containing a callsign to a destination buffer.
- */
-int parse_callsign(char *pSign, char *callBuf)
-{
-        char *inBuf=pSign;
-        int signLen=0;
-
-        while( (signLen < 10) && (isalnum(*inBuf) || *inBuf == '-') ) {
-                callBuf[signLen] = *inBuf;
-                signLen++;
-                inBuf++;
-        }
-        callBuf[signLen] = '\0';
-        return( strlen(callBuf) );
-}
-
-/*
- * handle the third party packets
- */
-void webdisplay_thirdparty(struct state *state, fap_packet_t *fap)
-{
-        char buf[512];
-        char fapbuf[512];
-        char tmpbuf[16];
-        char typebuf[16];
-        char *nextchar;
-        char *dst_callsign;
-        fap_packet_t *encap_fap = NULL;
-
-
-        /* Check for a third party packet */
-        if( ((nextchar=strchr(fap->orig_packet, ':')) != NULL)
-            && *(nextchar+1) == '}' ) {
-                strcpy(fapbuf, (nextchar+2));
-        } else {
-                printf("%s: Failed to verify encapsultated packet %s\n",
-                       __FUNCTION__, fap->orig_packet);
-                display_fap_pkt("third party verify failed", fap);
-                return;
-        }
-
-        /* Validated a 3rd party message packet, now parse it */
-        encap_fap=dan_parseaprs(fapbuf, strlen(fapbuf), 0);
-
-        if(encap_fap == NULL) {
-                printf("%s: Failed to parse encapsultated packet %s\n",
-                       __FUNCTION__, fap->orig_packet);
-                display_fap_pkt("third party parse failed", fap);
-                return;
-        }
-
-        if (encap_fap->error_code != NULL) {
-                display_fap_error("Third Party", state, encap_fap);
-                fap_free_wrapper("Third party", state, encap_fap);
-                return;
-        }
-
-        if (fap->type == NULL) {
-                strcpy(typebuf, "undefined");
-        } else {
-                sprintf(typebuf, "%02x", *fap->type);
-        }
-        printf("%s: successfully parsed an encapsultated packet type = %s\n",
-               __FUNCTION__, typebuf );
-        display_fap_pkt("third party parse ok", encap_fap);
-
-        state->stats.encapCount++;
-
-        /* Check for a third party packet that's a message */
-        if(encap_fap->message) {
-                display_fap_message(state, encap_fap);   /* DEBUG only - look at message packets */
-                if(isnewmessage(state, encap_fap)) {
-                        save_message(state, encap_fap);
-                        webdisplay_message(state, encap_fap);
-                }
-        } else {
-
-                if(encap_fap->src_callsign == NULL) {
-                        asprintf(&encap_fap->src_callsign, "NULL");
-                }
-
-                dst_callsign = encap_fap->destination;
-                if(encap_fap->destination == NULL) {
-                        if(encap_fap->dst_callsign != NULL) {
-                                dst_callsign = encap_fap->dst_callsign;
-                        } else if( (nextchar=strchr(encap_fap->orig_packet, '>')) != NULL) {
-                                if(parse_callsign(nextchar+1, tmpbuf) > 0) {
-                                        dst_callsign = tmpbuf;
-                                }
-                        }  else {
-                                asprintf(&dst_callsign, "NULL");
-                        }
-                }
-
-                if(encap_fap->orig_packet == NULL) {
-                        asprintf(&encap_fap->orig_packet, "NULL");
-                }
-
-#if 0
-                store_packet(state, fap);
-#endif
-
-                snprintf(buf, sizeof(buf), "%s->%s@%s: type: %s %s",
-                         encap_fap->src_callsign,
-                         dst_callsign,
-                         format_msgtime(encap_fap->timestamp),
-                         typebuf,
-                         encap_fap->orig_packet);
-
-                printf("Processed thirdparty: %s\n", buf);fflush(stdout);
-
-                _ui_send(state, "DB_PACKET", buf);
-                fap_free_wrapper("Third Party", state, encap_fap);
-        }
-}
-
-void webdisplay_unhandled(struct state *state, fap_packet_t *fap)
-{
-        char buf[512];
-
-        printf("%s: Call to: %s, Msg to: %s, pkt: %s\n",
-               __FUNCTION__, fap->dst_callsign, fap->destination, fap->orig_packet);
-
-        snprintf(buf, sizeof(buf), "%s->%s@%s: type: %02x %s",
-                 fap->src_callsign,
-                 fap->destination,
-                 format_msgtime(fap->timestamp),
-                 *fap->type,
-                 fap->orig_packet);
-        _ui_send(state, "DB_PACKET", buf);
-}
-
-#endif /* WEBAPP */
 void display_statistics(struct state *state)
 {
         json_object *jstats;
@@ -1429,23 +1072,33 @@ void parse_incoming_packet(struct state *state, char *packet, int len, int isax2
 
         fap = dan_parseaprs(packet, len, isax25);
 
-        if(fap->type != NULL) {
-                /* DEBUG: temporarily only filter message packets */
-                if ( (*fap->type == fapMESSAGE) ||
-                     fap->message != NULL) {
-                        printf("\n%s: %lu, type: 0x%02x -> %s\n",
-                               time2str(fap->timestamp, 0),
-                               state->stats.inPktCount,
-                               *fap->type, packet);
-                }
-        } else {
-                if(state->debug.console_display_filter & CONSOLE_DISPLAY_DEBUG) {
-                        printf("%s: %s FAP unhandled packet type\n%s\n",
-                               __FUNCTION__, time2str(fap->timestamp, 0), packet);
-                }
-        }
-
         if (!fap->error_code) {
+                if (fap->type != NULL) {
+                        /* Handle message packets */
+                        switch(*fap->type) {
+                                case fapMESSAGE:
+                                        handle_aprsMessages(state, fap, packet);
+                                        break;
+
+                                case fapTHIRD_PARTY:
+                                        handle_thirdparty(state, fap);
+                                        break;
+
+                                default:
+                                        if(fap->message != NULL) {
+                                                printf("%s: %s FAP error, message pointer set but type not fapMessage, type=0x%0x, message: %s\n",
+                                                       time2str(fap->timestamp, 0), __FUNCTION__,  *fap->type,  packet);
+                                        }
+                                        break;
+                        }
+
+                } else {
+                        if(state->debug.console_display_filter & CONSOLE_DISPLAY_DEBUG) {
+                                printf("%s %s: FAP unhandled packet type\n%s\n",
+                                        time2str(fap->timestamp, 0), __FUNCTION__, packet);
+                        }
+                }
+
                 store_packet(state, fap);
                 if (STREQ(fap->src_callsign, state->mycall)) {
                         state->digi_quality |= 1;
@@ -1464,7 +1117,6 @@ void parse_incoming_packet(struct state *state, char *packet, int len, int isax2
                 fap_free_wrapper("incoming pkt", state, fap);
         }
         display_statistics(state);
-
 }
 
 int handle_incoming_packet(struct state *state)
@@ -1557,6 +1209,9 @@ int display_gps_info(struct state *state)
         return 0;
 }
 
+/*
+ * Set system time using GPS
+ */
 int set_time(struct state *state)
 {
         struct posit *mypos = MYPOS(state);
@@ -1572,10 +1227,11 @@ int set_time(struct state *state)
         else if (mypos->sats < 3) {
                 return 1; /* Not enough sats, don't set */
         }
+
         /* Only set the Computer system time at boot up & once a day
          * after boot.
          * Setting system time frequently (every 2 minutes) causes
-         * programs that check & verify time like Dovecot to indciate a
+         * programs that check & verify time like Dovecot to indicate a
          * variation in system time of more than 7 seconds causing bad
          * things to happen.
          */
@@ -2254,6 +1910,9 @@ int should_beacon(struct state *state)
         return delta > req;
 }
 
+/*
+ * return true if beacon was sent successfully
+ */
 int beacon(struct state *state)
 {
         char *packet;
@@ -2288,16 +1947,16 @@ int beacon(struct state *state)
                 ret = send_beacon(state, packet);
                 free(packet);
         }
+        if(ret) {
+                state->last_beacon = time(NULL);
+                state->digi_quality <<= 1;
+                update_mybeacon_status(state);
 
-        state->last_beacon = time(NULL);
-        state->digi_quality <<= 1;
-        update_mybeacon_status(state);
+                _ui_send(state, "I_TX", "1000");
 
-        _ui_send(state, "I_TX", "1000");
-
-        state->last_beacon_pos = state->mypos[state->mypos_idx];
-
-        return 0;
+                state->last_beacon_pos = state->mypos[state->mypos_idx];
+        }
+        return ret;
 }
 
 int redir_log(struct state *state)
@@ -2424,7 +2083,7 @@ int canned_packets(struct state *state)
         while(bytes_read > 0) {
                 if ((bytes_read = fread(&recpkt, sizeof(recpkt_t), 1, fsrecpkt)) != 1) {
                         if(bytes_read == 0) {
-                                printf("Finished reading all %d packets, %d bytes\n",
+                                printf("Finished reading all %d packets, %zu bytes\n",
                                        cnt_pkts, cnt_bytes + cnt_pkts*sizeof(recpkt_t));
                                 break;
                         }
